third_party/WebKit/LayoutTests/http/tests/serviceworker/indexeddb.html

<!DOCTYPE html>
<title>Service Worker: Indexed DB</title>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="resources/test-helpers.js"></script>
<script>
var test = async_test('Verify Indexed DB operation in a Service Worker');
test.step(function() {
    var scope = 'resources/blank.html';
    navigator.serviceWorker.unregister(scope).then(
        doTest,
        unreached_rejection(test, 'Unregister should not fail')
    );

    function doTest() {
      navigator.serviceWorker.register('resources/indexeddb-worker.js',
                                      {scope: scope}).then(
          test.step_func(function(worker) {
              var messageChannel = new MessageChannel();
              messageChannel.port1.onmessage = test.step_func(onMessage);

              worker.postMessage({port: messageChannel.port2}, [messageChannel.port2]);
          }),
          unreached_rejection(test, 'Registration should succeed, but failed'));
    }

    function onMessage() {
        var openRequest = indexedDB.open('db');
        openRequest.onsuccess = test.step_func(function() {
            var db = openRequest.result;
            var tx = db.transaction('store');
            var store = tx.objectStore('store');
            var getRequest = store.get('key');
            getRequest.onsuccess = test.step_func(function() {
                assert_equals(getRequest.result, 'value',
                              'The get() result should match what the worker put().');
                test.done();
            });
        });
    }
});
</script>
